+++ .gitignore
@@ -1,0 +1,5 @@
+*/debug/*
+*/bin/*
+.vs/*
+*/x64/*
+*.swp
