--- convex_hull_cpp/convex_hull_cpp.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30320.27
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "convex_hull_cpp", "convex_hull_cpp\convex_hull_cpp.vcxproj", "{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Debug|x64.ActiveCfg = Debug|x64
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Debug|x64.Build.0 = Debug|x64
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Debug|x86.ActiveCfg = Debug|Win32
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Debug|x86.Build.0 = Debug|Win32
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Release|x64.ActiveCfg = Release|x64
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Release|x64.Build.0 = Release|x64
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Release|x86.ActiveCfg = Release|Win32
-		{CB82DDD8-37E8-44BC-A57E-B1D0BCC8569F}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {34A759D1-D946-4FF4-A95C-56F9263E7AC1}
-	EndGlobalSection
-EndGlobal
