--- convex_hull_cpp/convex_hull_cpp/Debug/convex_hull_cpp.tlog/convex_hull_cpp.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.26.28801:TargetPlatformVersion=10.0.18362.0:
-Debug|Win32|C:\Users\jk121\Desktop\codeing_study_C++\convex_hull_\convex_hull_cpp\|
